legacy/requirements.txt
Django
djangorestframework
flet
requests
sounddevice
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
numpy
soundfile
SQLAlchemy
